requirements.txt
azure-functions==1.21.3
azure-storage-blob==12.22.0
azure-ai-documentintelligence==1.0.0
azure-core==1.32.0
openai==1.57.2
requests==2.32.3
